Create a faketable object

Create a faketable object

Usage

faketable(data, table_def, rowId = NULL, show_delete = NULL)

Arguments

data

A data.frame

table_def

A table_def() object

rowId

A character vector of length one identifying which column is a primary key, if any.

show_delete

A named list passed to the shiny::column() that holds the Delete button. If NULL, the delete column will not be shown.

Value

A faketable object with the following properties:

data: The current state of the table of inputs

inserted: Rows from data that were inserted into the data (were not present in .raw_data)

updated: Rows from data that have been modified, but were present in .raw_data

deleted: Rows that were in .raw_data, but have been removed from and do not appear in data

.raw_data: The data originally pased to faketable() with the addition of a .rowId column as the first column. This column is calculated using by hashing either the provided rowId column or using the row number and system time.

.rowId: The value of the rowId argument

.deleted: All rows that have been removed from data, including those that were inserted then deleted

.table_def: A copy of the user supplied table_def() passed as the argument table_def

.show_delete: A copy of the user supplied list passed as the argument show_delete

Details

A faketable object is an S7::S7_object() with the class faketable. S7 object properties are accessed using an @, rather than the traditional $. For example, the property data for a faketable object called faketable can be accessed using faketable@data.

See also