insert, update, or delete rows from a faketables object

insert, update, or delete rows from a faketables object

Usage

insert(f_tab, x)

update(f_tab, x)

delete(f_tab, x)

Arguments

f_tab

A faketable() object

x

insert: A data.frame to add to the data in the faketable() object. If it does not already have a primary key column as specified in table_def(), one will be created and primary keys will be generated.

update: A data.frame with a primary key column as specified in table_def() with primary key values already present in the data.

delete: Either a character vector of values from the primary key column as specified in table_def() or a data.frame of rows to remove with that vector as a column.

Value

A faketable() object

Examples

if (FALSE) {
# insert
# to insert new rows of data from a data.frame called `ins`
f_tab <- faketablesServer(faketable = insert(f_tab(), ins))

# update
# to update with a new column called 'new' that has the value 'new'
faketable <- update(faketable, dplyr::mutate(faketable@x, 'new' = 'new'))

# delete
# to delete the first six rows of the data where the primary key column is `rowId`
rows_to_delete <- utils::head(faketable@x)
faketable <- delete(faketable, rows_to_delete$.rowId)
# OR
faketable <- delete(faketable, rows_to_delete)
}